booking-services/requirements.txt
motor
fastapi
datetime
pydantic
python-dotenv
uvicorn
python-multipart
python-jose